delete DateAccordion
